requirements.txt
keras==2.15.0
matplotlib==3.7.2
numpy==1.24.3
PyYAML==6.0.1
scipy==1.11.4
tensorflow==2.15.0
tensorflow_intel==2.15.0
tqdm==4.66.1
